fix dashboard sesuai spreadsheet po

+++ database/migrations/2023_11_24_024047_create_production_line_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('production_line', function (Blueprint $table) {
             $table->id();
             $table->string('nama_pl');
-            $table->integer('kapasitas_pl');
-            $table->timestamps();
+            $table->double('kapasitas_pl');
+            $table->timestamps();                                                               
         });
     }
